chore: refactor api types Mentions: #84
