+++ client/src/hooks/api/useGetUser.tsx
@@ -1,8 +1,9 @@
+import { IUser } from "api/types";
 import { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "store";
 import { setUser } from "store/slices/user";
-import axios, { AxiosResponse, IUser } from "../../api";
+import axios, { AxiosResponse } from "../../api";
 
 const useGetUser = () => {
   const authState = useSelector((state: RootState) => state.auth);
