+++ client/src/store/slices/exercises.ts
@@ -1,4 +1,4 @@
-import { IExercise } from "api";
+import { IExercise } from "api/types";
 
 export interface ExercisesState {
   exercises: IExercise[];
